+++ services/director-v2/src/simcore_service_director_v2/modules/api_keys_manager.py
@@ -16,7 +16,7 @@ from servicelib.redis import RedisClientSDK
 from settings_library.redis import RedisDatabase
 
 from ..core.settings import AppSettings
-from ..utils.distributed_identifier import BaseDistributedIdentifierManager
+from ..utils.base_distributed_identifier import BaseDistributedIdentifierManager
 from .rabbitmq import get_rabbitmq_rpc_client
 
 _CLEANUP_INTERVAL = timedelta(minutes=5)
