+++ services/director-v2/tests/unit/test_utils_distributed_identifier.py
@@ -15,7 +15,7 @@ from pytest_mock import MockerFixture
 from servicelib.redis import RedisClientSDK
 from servicelib.utils import logged_gather
 from settings_library.redis import RedisDatabase, RedisSettings
-from simcore_service_director_v2.utils.distributed_identifier import (
+from simcore_service_director_v2.utils.base_distributed_identifier import (
     BaseDistributedIdentifierManager,
 )
